LayoutTests/fast/forms/textarea-initial-caret-position.html

<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<div>
<p>This tests the initial caret position for textareas.</p>
<p>Bug 11746: REGRESSION(r14931): Outlook Web Access incorrectly positions the insertion point when replying to e-mail<br>
<a href="http://bugs.webkit.org/show_bug.cgi?id=11746">http://bugs.webkit.org/show_bug.cgi?id=11746</a></p>
</div>
<div>
<textarea id="ta" rows="5" cols="10">
1
2
3
4
5
6
7
8
9
10
</textarea>
</div>
<div id="console"></div>
<script type="text/javascript">
    var ta = document.getElementById('ta');
    ta.focus();
    shouldBe('ta.selectionStart', '0');
    shouldBe('ta.selectionEnd', '0');
</script>
<script type="text/javascript" src="../js/resources/js-test-post.js"></script>
</body>
</html>
